+++ Packages/com.unity.inputsystem/InputSystem/InputManager.cs
@@ -1450,8 +1450,10 @@ namespace UnityEngine.Experimental.Input
             //interactions.AddTypeRegistration("Swipe", typeof(SwipeInteraction));
 
             // Register composites.
+            composites.AddTypeRegistration("1DAxis", typeof(AxisComposite));
+            composites.AddTypeRegistration("2DVector", typeof(Vector2Composite));
             composites.AddTypeRegistration("Axis", typeof(AxisComposite));
-            composites.AddTypeRegistration("Dpad", typeof(DpadComposite));
+            composites.AddTypeRegistration("Dpad", typeof(Vector2Composite));
         }
 
         internal void InstallRuntime(IInputRuntime runtime)
@@ -1617,10 +1619,7 @@ namespace UnityEngine.Experimental.Input
             public StateChangeMonitorListener[] listeners;
             public DynamicBitfield signalled;
 
-            public int count
-            {
-                get { return signalled.length; }
-            }
+            public int count => signalled.length;
 
             public void Add(InputControl control, IInputStateChangeMonitor monitor, long monitorIndex)
             {
@@ -1705,7 +1704,7 @@ namespace UnityEngine.Experimental.Input
             }
         }
 
-        private void ResetControlPathsRecursive(InputControl control)
+        private static void ResetControlPathsRecursive(InputControl control)
         {
             control.m_Path = null;